Cache is_bfv check in ResourceTracker
Whether or not an instance is volume-backed shouldn't
change for the lifetime of the server, so rather than
figure that out during every operation (or periodic task)
for an instance in the resource tracker, get the is_bfv
value once and cache it. The cache entries are removed
when the instance is removed from the host (either deleted,
migrated or shelved offloaded).

Change-Id: Ic5188d4c5a7e8aa4c79c37e63124fe3a305ef568
Related-Bug: #1782393