Hyper-V + OVS: plug vifs before starting VMs
At the moment, the Hyper-V driver will bind OVS ports only after
starting the VMs. This is because of an old limitation of the OVS
Windows port, which has been addressed in OVS 2.5.

This also means that the Nova driver fails when waiting for neutron
vif plug events, since this step is performed later. So deployers
are forced to set "vif_plugging_is_fatal", which means that the
ports will be down, at least for a short time after the instance is
reported as ACTIVE. This also breaks some Tempest tests.

Since the Windows OVS version has been fixed, we can now bind the
ports before actually starting the VMs. Note that this fix has been
merged in the out-of-tree compute-hyperv some time ago.

Co-authored-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>

Change-Id: I0116a593c29808a53423cc9bbe098f2661fcb584
Closes-Bug: #1684250