Use source vifs when unplugging on source during post live migrate
With the changes in blueprint neutron-new-port-binding-api, we now
manage port bindings on both the source and destination host and
those can be different vif types to support live migrating between
different network backends, e.g. linuxbridge-ovs, ovs->ovn, etc.

The dest host port bindings are created during pre_live_migration
and then activated on the source host during post-copy (for the
libvirt driver) or in _post_live_migration when we call the
migrate_instance_start() method in the network API. When the dest
host port bindings are activated, the source host port bindings
are automatically deactivated (it's an atomic operation in neutron).

Depending on when we refresh the nw info cache on the instance,
we can end up calling driver.post_live_migration_at_source() with
information about the unbound source port binding, or the vif type
for the destination host port binding. Because of that, unplug_vifs
on the source can fail since it's not using the proper vif type.

To resolve that, we use the stashed VIFMigrateData.source_vif
when unplugging vifs on the source host. Those source vifs are
an exact copy of the vifs on the instance before live migration
started on the source host, so they are appropriate for unplugging
on the source during _post_live_migration.

Change-Id: Ie6439499cc7c6f05eec692dda677caf649dd7c94
Closes-Bug: #1783917