Add functional test for forced live migration rollback allocs
Per bug 1784022, it was not immediately clear that on a forced
host live migration failure that the allocations on the dest
host would be deleted and the original allocations for the instance
would be back on the source host, but it is indeed working but
as far as I could tell we didn't have a functional test for it,
only the non-forced case. So this change adds the force=True
wrinkle to the same existing functional test and also adds
assertions on where the allocations exist for the source and dest
nodes and the migration and instance records at the point of
failure.

Change-Id: Iaef9a4456c9c2bd9b2f4257b4ddd3efa3dea50f6
Related-Bug: #1784022