Create request spec, build request and mappings in one transaction
The transaction context is currently not shared when creating the
RequestSpec, BuildRequest and InstanceMapping.  Because of this,
it is possible that the database ends in an inconsistent state
due to the fact that one of these was created and the system
crashed afterwards.

This patch adds a function which handles the creation of all those
resources in a single transaction.

Co-Authored-By: melanie witt <melwittt@gmail.com>

Closes-Bug: #1784093

Change-Id: If897a0d721180152ebdceb7a0c23e8f283ce6d10
(cherry picked from commit 85f8d033d27b31a6398529e0a25da74eae523b08)