Hook resource_tracker to remove stale node information
When we remove a node from a host due to rebalance or decommission,
we need to tell the resource tracker so it stops keeping track of
the details for that node. This adds a hook for doing that and
calls it from the place in compute manager where we purge the
record from the database.

Change-Id: Ie6b6bb2a9e8a4ad33675fccb3827e8197fa16398
Closes-Bug: #1784874
(cherry picked from commit 99db9faae5c1e84e5aa8586c32ffcca7a61ae276)
(cherry picked from commit 7302a625ee9af2c9dd88f8630f1e95fc5977a70a)