Grease test_try_deallocate_network_retry_direct
That's "grease" in the sense of "lubricate", or "make faster". Not in
the sense of "kill with an M3 submachine gun".

This test hits the compute manager's _try_deallocate_network helper in
such a way as to deliberately activate and exhaust its RetryDecorator,
which adds 12s to the test time. This patch mocks the sleep in that
retry loop, taking the test down to sub-second like it should be.

Change-Id: I60d11fc9a9e8569b1663c7319a5c25b921c5de1a