xx_instance_type_id in list_migrations should be integer
new_instance_type_id and old_instance_type_id in response
of list_migrations should be integer, not string.

Besides, this flavor id is not a common one specified by
users, but an internal one, so this is also to add a NOTE
to avoid the confusion.

"old_instance_type_id": 5140

https://developer.openstack.org/api-ref/compute/#list-migrations

Change-Id: I43c454ac1d7f7ef413d44fe411459e989611cadb