Fix image-defined numa claims during evacuate
When evacuating, the API does not send the image_ref
to the compute so currently the compute manager
rebuild_instance() method will just pass an empty
dict for image_meta to the rebuild_claim, which means
if the server was originally created with an image that
has numa-related constraints, like hw_numa_nodes, those
constraints would not be applied to the destination host
during the evacuate.

This change simply checks for evacuate if image_ref is
not provided and pulls the image_meta off the instance
which was stashed in the instance.system_metadata during
server create (see get_system_metadata_from_image usage
in the compute API).

This fix was ported from the starlingx-staging/stx-nova
repo commit 71acfeae0.

Change-Id: If548fa3436174b1eae08cdcf6578020cc0c7b81f
Closes-Bug: #1785318