Fix resize revert to use non-legacy alloc handling
During the revert of resize or migrate nova tries to drop the
allocation on the target host. But first it decides if the resize uses
the new style allocation handling where the source host allocation is
held by the migration. If not then the legacy double-up allocation
handling is triggered.

However the implementation tried to check if the migration holds
allocation on the destination host and that is clearly wrong as migration
only holds alloction on the source host.

This patch now changes the check to simply assert if there is any
allocation held by the migration.uuid.

Change-Id: I41c10ca5caa01dda950dd17319889963850fed59
Closes-Bug: #1785776