Minimal construct plumbing for nova service-list when a cell is down
This patch modifies the existing behavior of nova service-list which
ignores a down cell to returning a partial construct consisting of
the binary and host info obtained from the API host_mappings table
for the compute services in the down cells. This is enabled by passing
a new enough microversion, which is added in a later patch in this
series.

Related to blueprint handling-down-cell

Change-Id: I9cf5e4eb7f70c495001fd064c060d7d5b3dd5bff