Consumer gen: remove_provider_from_instance_allocation
This patch changes the scheduler report client
remove_provider_from_instance_allocation function to use placement API
microversion 1.28 that includes consumer generation handling. If
placement returns 409 conflict remove_provider_from_instance_allocation
returns False in the same way as in case of other non 204 results.

There are plenty of places in nova calling
remove_provider_from_instance_allocation:
* dropping allocation on the destination of a live migrate if pre check
  fails
* compute host startup deletes allocation of an instance that was
  successfully evacuated from the host while it was down
* rebuild fails on the destination of an evacuation and therefore
  the allocation on the destination is deleted
* pre resize fails therefore allocation on the destination is deleted

This patch does not change the external behavior of
remove_provider_from_instance_allocation as it still returns False in
caes of failure. Therefore this patch does not introduce new functional
test for the above cases but rely on the existing coverage. However based
on the other patches in this series the error handling in these cases
is suspected to be imperfect.

Blueprint: use-nested-allocation-candidates
Change-Id: I328a18a723d0f593eea491f788a6e256d6e0c127