Py3 fix in fake image service
In py3, a file opened with 'b' must be written with a bytes object, not
a string (this restriction does not exist in py2):

 >>> f = open('/tmp/foo', 'wb+')
 >>> f.write('foo')
 Traceback (most recent call last):
   File "<stdin>", line 1, in <module>
 TypeError: a bytes-like object is required, not 'str'

The fake image service's download method was writing '' by default when
no image data was present, which would result in the above exception if
we ever hit that code path. This fix changes it to b''.

Change-Id: Id4644f611fd48936b9ef036c18174dba0013c84a