Add an example to add more pci devices in nova.conf
The help for the [pci]aliases config option was confusing in that it
wasn't clear how to specify multiple aliases. The default value being []
made it seem like you ought to be able to specify a list as:

 alias = [{...}, {...}, ...]

This won't actually work; you instead have to say:

 alias = {...}
 alias = {...}
 ...

So this patch adds a multi-alias example to the help string and
explicitly calls out the fact that you can't use list values.

See the referenced bug for more details.
Change-Id: I71d4c440340db31d22a73a8b3e3abc51771aa7bc
Closes-Bug: #1786870