Make monkey patch work in uWSGI mode
There was a eventlet.monkey_patch [1] when we launch a nova
process(like nova-api, nova-compute), but it's invalid under
the uwsgi mode.

But there are concurrency requirements in the api service, such
as, when listing instances cross multiple cells we're using
greenthreads and oslo.db does a time.sleep to allow switching
greenthreads [2].

So, in this patch we add the monkey_patch in the uwsgi
application setup and refactor the monkey patching to use common
code.

[1] https://github.com/openstack/nova/blob/233ea58/nova/cmd/__init__.py#L26
[2] https://github.com/openstack/oslo.db/blob/9c66959/oslo_db/sqlalchemy/engines.py#L51

Closes-bug: #1787331

Change-Id: Ie7bf5d012e2ccbcd63c262ddaf739782afcdaf56