List instances from all cells explicitly
This is a minor refactor with no functional change that will make
the next patch easier. If we're not limiting cell scope during an
instance list, we just let context.scatter_gather_all_cells() load
and use the global list of cells. Since we will want to count
those cells and adjust the batch size, we might as well just always
pass in the list of cells, which will either be a subset or all
of them.

Change-Id: Id167ab6d01698878d783999d54bcf77f37eec020