Filter out instances without a host when populating AZ
It could happen that instance does not have a host set, e.g.
when its creation failed before it was scheduled.
During online_migration, populate_missing_availability_zones tries to
add missing AZs to all instances. However for instances without a host
there's no reasonable value for AZ (we can't use a logic that bases
the value on a host) so let's skip this kind of instances completely.

Change-Id: Ic6060beaa08af5ea70e5e54fffb94eea58aa7bbf
Closes-Bug: #1788115