libvirt: Don't react to VIR_DOMAIN_EVENT_SUSPENDED_MIGRATED events
The VIR_DOMAIN_EVENT_SUSPENDED_MIGRATED event is sent when live
migration completes but does not indicate if the migration was
successful or a failure, for that we need to check the actual
job status.

For now, to fix the regression introduced in change
Ic5cab99944df9e501ba2032eb96911c36304494d in Rocky and mitigate
the risk of a change in the Rocky GA, we just don't deal with
that event. In the case of live migration failing, things are the
same as before Rocky. In the case of live migration success, we
activate the destination host port bindings in _post_live_migration
so the network downtime benefit is not realized in that case, but
we also don't mess up the network with incorrectly activating the
dest host port bindings in ComputeManager.handle_lifecycle_event.
If post-copy is in effect, we still get the reduced network downtime
benefit from blueprint neutron-new-port-binding-api.

This is a partial fix and the full fix will include checking the
job status when we get VIR_DOMAIN_EVENT_SUSPENDED_MIGRATED but is
likely too risky of a change for the Rocky GA.

Change-Id: If6779b633310c46c21bf7b3b3769d17e486f3945
Partial-Fix: #1788014