Make RecordWrapper record RequestContext and expose cell_uuid
This makes it easier to tell which cell a result came from when
we are merging results from multiple cells. Later in this set,
we will be processing results from cells inline, instead of
waiting for all cells to complete before processing the results.
That means in order to know which cells did and did not reply,
we need to be able to determine that from each result on the
other side of heapq.merge(). Since we already have a record
wrapper class, it is trivial to stash that information there.

Change-Id: I0631aad1f4d2ce8b786cae74de16d7dd6e3b1411